fix(chat): fix issues with idebars overlay, bottom panel highlight)
