+++ apps/chat/src/components/Chat/Chat.tsx
@@ -1,4 +1,3 @@
-import { FloatingOverlay } from '@floating-ui/react';
 import { IconPlayerPlay } from '@tabler/icons-react';
 import {
   memo,
@@ -36,7 +35,7 @@ import {
   ConversationsTemporarySettings,
   MergedMessages,
 } from '@/src/types/chat';
-import { EntityType, ScreenState } from '@/src/types/common';
+import { EntityType } from '@/src/types/common';
 import { Translation } from '@/src/types/translation';
 
 import { AddonsSelectors } from '@/src/store/addons/addons.reducers';
@@ -136,7 +135,6 @@ const ChatView = memo(() => {
   const talkToConversationId = useAppSelector(
     ConversationsSelectors.selectTalkToConversationId,
   );
-  const isAnyMenuOpen = useAppSelector(UISelectors.selectIsAnyMenuOpen);
   const isIsolatedView = useAppSelector(SettingsSelectors.selectIsIsolatedView);
   const installedModelIds = useAppSelector(
     ModelsSelectors.selectInstalledModelIds,
@@ -159,8 +157,6 @@ const ChatView = memo(() => {
   const [isWideLayout, setIsWideLayout] = useState(
     checkIsWideLayout(mergedMessages.length, isCompareMode),
   );
-
-  const screenState = useScreenState();
 
   const handleTalkToConversationId = useCallback(
     (conversationId: string | null) => {
@@ -532,10 +528,6 @@ const ChatView = memo(() => {
     !messageIsStreaming &&
     !isLastMessageError &&
     !notAvailableEntityType;
-  const showFloatingOverlay =
-    (screenState === ScreenState.SM || screenState === ScreenState.MD) &&
-    isAnyMenuOpen &&
-    !isIsolatedView;
   const isModelsInstalled = selectedConversations.every((conv) =>
     installedModelIds.has(conv.model.id),
   );
@@ -603,7 +595,6 @@ const ChatView = memo(() => {
       data-qa="chat"
       id="chat"
     >
-      {showFloatingOverlay && <FloatingOverlay className="z-30 bg-blackout" />}
       {modelError ? (
         <ErrorMessageDiv error={modelError} />
       ) : customViewer ? (
