+++ apps/chat/src/components/NavigationWrapper.tsx
@@ -67,11 +67,9 @@ const NavigationButton = ({
       data-qa={dataQa}
       onClick={onClick}
       className={classNames(
-        'flex max-h-[52px] min-w-[72px] shrink-0 cursor-pointer select-none flex-col items-center justify-center gap-[2px] rounded border border-transparent transition-colors duration-200 hover:bg-accent-primary-alpha hover:disabled:bg-transparent md:min-w-min md:p-[9px]',
-        {
-          'rounded-full': rounded,
-          '!border-accent-primary': rounded && selected,
-        },
+        'flex max-h-[52px] min-w-[72px] shrink-0 cursor-pointer select-none flex-col items-center justify-center gap-[2px] rounded border border-transparent transition-colors duration-200 active:bg-accent-primary-alpha active:disabled:bg-transparent md:min-w-min md:p-[9px] md:hover:bg-accent-primary-alpha md:active:bg-transparent md:hover:disabled:bg-transparent md:active:disabled:bg-transparent',
+        rounded && 'rounded-full',
+        rounded && selected && '!border-accent-primary',
       )}
     >
       {Icon && (
