+++ apps/chat/src/pages/index.tsx
@@ -1,4 +1,10 @@
+import { FloatingOverlay } from '@floating-ui/react';
+
+import { useScreenState } from '../hooks/useScreenState';
+
 import { getCommonPageProps } from '@/src/utils/server/get-common-page-props';
+
+import { ScreenState } from '../types/common';
 
 import { MigrationSelectors } from '../store/migration/migration.reducers';
 import { useAppSelector } from '@/src/store/hooks';
@@ -47,6 +53,10 @@ function Home() {
   const showSelectToMigrateWindow = useAppSelector(
     UISelectors.selectShowSelectToMigrateWindow,
   );
+  const isAnyMenuOpen = useAppSelector(UISelectors.selectIsAnyMenuOpen);
+  const isIsolatedView = useAppSelector(SettingsSelectors.selectIsIsolatedView);
+
+  const screenState = useScreenState();
 
   if (conversationsToMigrateCount !== 0 || promptsToMigrateCount !== 0) {
     if (
@@ -56,6 +66,11 @@ function Home() {
       return window.location.reload();
     }
   }
+
+  const showFloatingOverlay =
+    (screenState === ScreenState.SM || screenState === ScreenState.MD) &&
+    isAnyMenuOpen &&
+    !isIsolatedView;
 
   return (
     <>
@@ -78,6 +93,9 @@ function Home() {
           {enabledFeatures.has(Feature.Header) && <Header />}
           <div className="flex w-full grow overflow-auto">
             <div className="flex min-w-0 grow flex-col">
+              {showFloatingOverlay && (
+                <FloatingOverlay className="z-30 bg-blackout" />
+              )}
               <AnnouncementsBanner />
               <Chat />
               <ImportExportLoader />
