+++ apps/chat/src/store/ui/ui.selectors.ts
@@ -69,7 +69,8 @@ export const selectIsAnyMenuOpen = createSelector(
   (state, enabledFeatures) =>
     (state.showPromptbar && enabledFeatures.has(Feature.PromptsSection)) ||
     (state.showChatbar && enabledFeatures.has(Feature.ConversationsSection)) ||
-    state.isProfileOpen,
+    state.isProfileOpen ||
+    state.showWidgetbar,
 );
 
 export const selectCollapsedSections = //TODO: review later how it is used
